feat(gpg_secretstore): asserting recipients
Instead of manually setting the reencrypt parameter, we're automatically
reencrypting secrets now if the list of recipients does not match. This
way, we can make sure that recipients are kept up to date here. We're
also logging differences in recipients, as well as adding more useful
messages to the return value in general.

In addition to that, the module has seen quite some refactoring, making
it a lot more maintainable.